pyproject.toml
[project]
name = "evalkit"
version = "0.1.0"
description = "Production-grade evaluation and optimization system for LLM-powered features"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "rich>=13.6.0",
    "pydantic>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.28.0",
    "faiss-cpu>=1.7.4",
    "qdrant-client>=1.6.0",
    "openai>=1.3.0",
    "prometheus-client>=0.17.0",
    "python-crontab>=3.0.0",
    "wandb>=0.15.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=evalkit" 